Added new timout output to picosat

+++ satfc/src/main/java/ca/ubc/cs/beta/stationpacking/solvers/sat/solvers/picoSAT/PicoSATSolver.java
@@ -199,6 +199,10 @@ public class PicoSATSolver extends AbstractCompressedSATSolver {
 
 
 
+            if (satResult == SATResult.TIMEOUT) {
+                log.info("returning timeout object");
+                return SATSolverResult.timeout(walltime);
+            }
 
             tempIn.delete();
             tempOutPico.delete();
